feat : BigInter -> int 변경, 단 건 조회 기능, partyType으로 필터링 조회기능 추가

+++ src/main/java/com/ma_recruit/dto/party/request/PartyPostCreateRequestDto.java
@@ -21,7 +21,7 @@ public class PartyPostCreateRequestDto {
     private Job job;
 
     @NotNull
-    private BigInteger mapId;
+    private int mapId;
 
     @Min(1)
     private int member_max_count;
