+++ src/main/java/com/ma_recruit/repository/guild/GuildPostRepository.java
@@ -1,13 +1,11 @@
 package com.ma_recruit.repository.guild;
 
 import com.ma_recruit.entity.guild.GuildPost;
-import com.ma_recruit.entity.party.PartyPost;
+import com.ma_recruit.entity.party.PartyType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface GuildPostRepository extends JpaRepository<GuildPost, BigInteger> {
+public interface GuildPostRepository extends JpaRepository<GuildPost, Integer> {
     Page<GuildPost> findAll(Pageable pageable);
 }
