+++ src/main/java/com/ma_recruit/repository/member/MemberRepository.java
@@ -5,10 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, BigInteger> {
+public interface MemberRepository extends JpaRepository<Member, Integer> {
     @Query("select m from Member m where m.discord_id = :discordId")
     Optional<Member> findByDiscordId(@Param("discordId") String discordId);
 }
